# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Presentation": {
        "additionalProperties": false,
        "properties": {
          "compact_explanations": {
            "default": false,
            "title": "Compact Explanations",
            "type": "boolean"
          },
          "font_scale": {
            "default": 1,
            "maximum": 1.5,
            "minimum": 1.0,
            "title": "Font Scale",
            "type": "number"
          },
          "reduced_motion": {
            "default": true,
            "title": "Reduced Motion",
            "type": "boolean"
          }
        },
        "title": "Presentation",
        "type": "object"
      },
...
      "ProblemPublic": {
        "properties": {
          "activity_state": {
            "default": "ready",
            "enum": [
              "generating",
              "reference_capture",
              "ready"
            ],
            "title": "Activity State",
            "type": "string"
          },
          "assistance_level": {
            "title": "Assistance Level",
            "type": "integer"
          },
          "concept_focus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Concept Focus"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "format_constraints": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format Constraints"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "operations": {
            "items": {
              "$ref": "#/components/schemas/OperationPublic"
            },
            "title": "Operations",
            "type": "array"
          },
          "problem_text": {
            "title": "Problem Text",
            "type": "string"
          },
          "reference_source": {
            "anyOf": [
              {
                "enum": [
                  "topic",
                  "reference_text",
                  "reference_photo"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Source"
          },
          "session_id": {
            "format": "uuid",
            "title": "Session Id",
            "type": "string"
          },
          "skill_id": {
            "title": "Skill Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "template_id": {
            "title": "Template Id",
            "type": "string"
          },
          "template_version": {
            "title": "Template Version",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "session_id",
          "template_id",
          "template_version",
          "skill_id",
          "problem_text",
          "format_constraints",
          "status",
          "created_at",
          "updated_at",
          "version",
          "assistance_level",
          "operations"
        ],
        "title": "ProblemPublic",
        "type": "object"
      },
...
      "ProfileSettings": {
        "additionalProperties": false,
        "properties": {
          "custom_instructions": {
            "default": "",
            "maxLength": 2000,
            "title": "Custom Instructions",
            "type": "string"
          },
          "difficulty": {
            "default": "standard",
            "enum": [
              "introductory",
              "standard",
              "challenge"
            ],
            "title": "Difficulty",
            "type": "string"
          },
          "hint_policy": {
            "const": "progressive",
            "default": "progressive",
            "title": "Hint Policy",
            "type": "string"
          },
          "language": {
            "const": "en",
            "default": "en",
            "title": "Language",
            "type": "string"
          },
          "name": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "presentation": {
            "$ref": "#/components/schemas/Presentation"
          },
          "question_pacing": {
            "const": 1,
            "default": 1,
            "title": "Question Pacing",
            "type": "integer"
          },
          "session_problem_limit": {
            "default": 5,
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Session Problem Limit",
            "type": "integer"
          },
          "solution_policy": {
            "default": "after_two_attempts",
            "enum": [
              "on_request",
              "after_two_attempts",
              "adult_only"
            ],
            "title": "Solution Policy",
            "type": "string"
          },
          "teaching_style": {
            "default": "guided",
            "enum": [
              "guided",
              "direct",
              "worked_example"
            ],
            "title": "Teaching Style",
            "type": "string"
          },
          "topics": {
            "items": {
              "type": "string"
            },
            "maxItems": 7,
            "minItems": 1,
            "title": "Topics",
            "type": "array"
          },
          "verbosity": {
            "default": "standard",
            "enum": [
              "brief",
              "standard",
              "detailed"
            ],
            "title": "Verbosity",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ProfileSettings",
        "type": "object"
      },
...
      "SessionInput": {
        "additionalProperties": false,
        "properties": {
          "learner_id": {
            "format": "uuid",
            "title": "Learner Id",
            "type": "string"
          },
          "profile_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Version Id"
          }
        },
        "required": [
          "learner_id"
        ],
        "title": "SessionInput",
        "type": "object"
      },
      "SessionPublic": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "learner_id": {
            "format": "uuid",
            "title": "Learner Id",
            "type": "string"
          },
          "problems": {
            "items": {
              "$ref": "#/components/schemas/ProblemPublic"
            },
            "title": "Problems",
            "type": "array"
          },
          "profile": {
            "$ref": "#/components/schemas/ProfileSettings"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "learner_id",
          "status",
          "profile",
          "problems"
        ],
        "title": "SessionPublic",
        "type": "object"
      },
...
      "SessionSummary": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "learner_id": {
            "format": "uuid",
            "title": "Learner Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "learner_id",
          "status",
          "created_at"
        ],
        "title": "SessionSummary",
        "type": "object"
      },
...
      "TutorActivityInput": {
        "additionalProperties": false,
        "properties": {
          "difficulty": {
            "anyOf": [
              {
                "enum": [
                  "introductory",
                  "standard",
                  "challenge"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty"
          },
          "reference_text": {
            "anyOf": [
              {
                "maxLength": 8000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Text"
          },
          "source": {
            "default": "topic",
            "enum": [
              "topic",
              "reference_text",
              "reference_photo"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "title": "TutorActivityInput",
        "type": "object"
      },
      "TutorSettingsInput": {
        "additionalProperties": false,
        "properties": {
          "difficulty": {
            "anyOf": [
              {
                "enum": [
                  "introductory",
                  "standard",
                  "challenge"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty"
          },
          "initiative": {
            "enum": [
              "tutor_led",
              "balanced",
              "learner_led"
            ],
            "title": "Initiative",
            "type": "string"
          }
        },
        "required": [
          "initiative"
        ],
        "title": "TutorSettingsInput",
        "type": "object"
      },
      "TutoringSessionInput": {
        "additionalProperties": false,
        "properties": {
          "difficulty": {
            "default": "standard",
            "enum": [
              "introductory",
              "standard",
              "challenge"
            ],
            "title": "Difficulty",
            "type": "string"
          },
          "initial_activity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TutorActivityInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "initiative": {
            "default": "balanced",
            "enum": [
              "tutor_led",
              "balanced",
              "learner_led"
            ],
            "title": "Initiative",
            "type": "string"
          },
          "learner_id": {
            "format": "uuid",
            "title": "Learner Id",
            "type": "string"
          },
          "topic": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "learner_id",
          "topic"
        ],
        "title": "TutoringSessionInput",
        "type": "object"
      },
      "TutoringSessionPublic": {
        "properties": {
          "difficulty": {
            "enum": [
              "introductory",
              "standard",
              "challenge"
            ],
            "title": "Difficulty",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "initiative": {
            "enum": [
              "tutor_led",
              "balanced",
              "learner_led"
            ],
            "title": "Initiative",
            "type": "string"
          },
          "learner_id": {
            "format": "uuid",
            "title": "Learner Id",
            "type": "string"
          },
          "problems": {
            "items": {
              "$ref": "#/components/schemas/ProblemPublic"
            },
            "title": "Problems",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "topic": {
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "id",
          "learner_id",
          "status",
          "topic",
          "initiative",
          "difficulty",
          "problems"
        ],
        "title": "TutoringSessionPublic",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Shepherd Academy Universe API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/sessions": {
      "get": {
        "operationId": "sessions_api_v1_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SessionSummary"
                  },
                  "title": "Response Sessions Api V1 Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Sessions",
        "tags": [
          "practice"
        ]
      },
      "post": {
        "operationId": "start_session_api_v1_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Session",
        "tags": [
          "practice"
        ]
      }
    },
    "/api/v1/sessions/{session_id}": {
      "get": {
        "operationId": "read_session_api_v1_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Session",
        "tags": [
          "practice"
        ]
      }
    },
...
    "/api/v1/tutor/sessions": {
      "get": {
        "operationId": "sessions_api_v1_tutor_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TutoringSessionPublic"
                  },
                  "title": "Response Sessions Api V1 Tutor Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Sessions",
        "tags": [
          "AI tutoring"
        ]
      },
      "post": {
        "operationId": "create_session_api_v1_tutor_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutoringSessionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutoringSessionPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "AI tutoring"
        ]
      }
    },
    "/api/v1/tutor/sessions/{session_id}": {
      "get": {
        "operationId": "read_session_api_v1_tutor_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutoringSessionPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Session",
        "tags": [
          "AI tutoring"
        ]
      }
    },
    "/api/v1/tutor/sessions/{session_id}/activities": {
      "post": {
        "operationId": "activity_api_v1_tutor_sessions__session_id__activities_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorActivityInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activity",
        "tags": [
          "AI tutoring"
        ]
      }
    },
    "/api/v1/tutor/sessions/{session_id}/settings": {
      "post": {
        "operationId": "settings_api_v1_tutor_sessions__session_id__settings_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorSettingsInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutoringSessionPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Settings",
        "tags": [
          "AI tutoring"
        ]
      }
    },
...
  }
}